backend/requirements.txt
fastapi==0.98.0
loguru==0.7.0
numpy==1.25.0
openai==0.27.8
pandas==2.0.3
pinecone_client==2.2.2
tqdm==4.65.0
uvicorn==0.22.0
mysql-connector-python==8.0.33
boto3==1.28.3
sendgrid==6.10.0